Import data

Imports .csv-files generated by QuantaSoft.

import_QS_files(
  paths,
  Ch1_is_mutation = TRUE,
  annotations = NULL,
  sample_annotations = NULL,
  merge_wells = "none",
  merge_files = FALSE
)

Arguments

paths
Character vector. Specifies which files and/or directories to load. For directories all .csv files within are loaded.

Ch1_is_mutation
Logical. Control if Ch1 is used to measure presence of mutant DNA (and Ch2 for wild type DNA). If this is FALSE Ch2 is used instead (and Ch1 for wild type).

annotations
Named list or data.frame. Annotations to be added to all samples.

sample_annotations
data.frame. Similar to annotations, but annotations are specific to each sample. This need to include a column "Sample", which will be used for joining.

merge_wells
String. Controls if wells from the same sample ("Sample") should be merged within a dataset (file). There are 4 options:

merge_wells="yes": Merge wells. Discards QS merged wells (e.g. "M01").

merge_wells="no": Do **not** merge wells. Discard QS merged wells (e.g. "M01").

merge_wells="qs": The merged wells from QuantaSoft (e.g. "M01") is used if these are present.

merge_wells="none": No merging is done.

Default if "none".

merge_files
Logical. If this and merge_wells is TRUE, samples across files are also merged. Default is FALSE.

Value

A data.frame with the columns:

FileName:

Name of the file imported.

Well:

Well identifier.

Sample:

Sample name from QuantaSoft.

Ch1TargetType:

The target type for channel 1.

Ch2TargetType:

The target type for channel 2.

Target:

Target name from QuantaSoft.

MutantOnlyDroplets:

The count of mutant only droplets.

WildtypeOnlyDroplets:

The count of wild type only droplets.

DoubleNegativeDroplets:

The count of double negative droplets.

DoublePositiveDroplets:

The count of double positive droplets.

TotalDroplets:

The total number of droplets.

MergedWells:

If merge_wells=TRUE this is a string of the wells merged. Otherwise this will be NA.

NumberOfMergedWells:

The number of wells merged.

See also